backend/requirements.txt
fastapi
uvicorn
pinecone-client
requests
python-dotenv
